backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
web3>=6.15.0
slowapi>=0.1.9
aiosqlite>=0.20.0
databases>=0.9.0
cryptography>=42.0.0
PyJWT[crypto]>=2.8.0
